+++ src/app/components/books/character-group/character-group.component.html
