+++ src/app/components/books/character-group/character-group.component.spec.ts
