+++ src/app/components/books/characters/characters.component.ts
