+++ src/app/components/clone-wars-chronological/clone-wars-chronological.component.scss
@@ -1,4 +1,4 @@
-@import "../../styles.scss";
+@import "../../../styles.scss";
 
 #chronological-episodes-grid {
   grid-template-columns: 221px 506px 1fr;
