+++ src/app/components/clone-wars-review/clone-wars-review.component.scss
@@ -1,4 +1,4 @@
-@import "../../styles.scss";
+@import "../../../styles.scss";
 
 #clone-wars-container {
   grid-template-columns: 1fr 1fr;
